Estandarización de código y naming conventions, agregado manejo de excepciones, estandarización de información de funciones, unificación de todos los módulos en uno solo, cambio de variables y funciones a español
